src/job/requirements.txt
# Cloud Run Job: Binance + BigQuery warm-up + pandas-ta + Pub/Sub (pinned for reproducibility)
# pandas>=2.1 for Python 3.12 wheels (2.0.x builds from source and can fail with pkg_resources)
requests~=2.28.0
pandas>=2.1,<3
pandas-ta>=0.3.14b0
google-cloud-bigquery~=3.14.0
google-cloud-pubsub~=2.18.0
# Soporte de tipos BigQuery -> pandas (para to_dataframe())
db-dtypes>=1.0.0,<2
